requirements.txt
robotframework==5.0.1
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.4
robotframework-seleniumlibrary==3.3.1